update session link
