+++ src/components/2025/nav.js
@@ -97,10 +97,10 @@ function Nav() {
               </div>
               <div className="bg-white mx-2 w-60 flex sm:flex-row flex-col justify-center sm:mb-16 md:mb-10 lg:mb-0">
                 <Link
-                  to="/speaker-schedule"
+                  to="/schedule"
                   className="font-clashDisplay font-bold whitespace-nowrap  p-5 text-lg tracking-wider mx-5 w-fit flex items-center"
                 >
-                  Speaker Schedule
+                  Schedule
                   <svg
                     className="ml-4"
                     width="14"
@@ -155,11 +155,11 @@ function Nav() {
       >
         <div className="mx-2 flex flex-col bg-white my-5">
           <Link
-            to="/speaker-schedule"
+            to="/schedule"
             className="font-clashDisplay font-bold  p-5 text-black tracking-wider whitespace-nowrap mx-5 flex items-center justify-center"
             onClick={() => toggleExpansion(!isExpanded)}
           >
-            Speaker Schedule &nbsp;
+            Schedule &nbsp;
             <svg width="14" height="15" viewBox="0 0 14 15" fill="none" xmlns="http://www.w3.org/2000/svg">
               <path
                 d="M0.75 13.75L13.25 1.25M13.25 1.25H3.875M13.25 1.25V10.625"
