+++ src/pages/schedule.js
@@ -2,24 +2,25 @@ import React from "react";
 
 import Layout from "../components/layout";
 import SEO from "../components/SEO/seo";
+import Nav from "../components/2025/nav";
+import Footer from "../components/2025/footer";
 
-import Nav from "../components/2023/nav";
-import Breadcrumb from "../components/2023/breadcrumb";
-import Sched from "../components/schedEmbed";
-import Footer from "../components/2023/footer";
-
-function Schedule() {
+function SpeakerSchedulePage() {
   return (
     <Layout>
-      <SEO title="Schedule" />
+      <SEO title="Speaker Schedule" />
       <Nav />
-      <Breadcrumb name="Schedule" />
-      <div className="pt-10 bg-white border-b-2">
-        <Sched slug="oscafest23" />
-      </div>
+      <iframe
+        src="https://sessionize.com/view/iwyk9p10/GridSmart?format=Embed_Styled_Html&isDark=False&title=Open%20Source%20Festival%202025"
+        width="100%"
+        height="800"
+        frameBorder="0"
+        title="Open Source Festival 2025 Schedule"
+        className="border-none -mt-[130px] lg:-mt-[90px]"
+      ></iframe>
       <Footer />
     </Layout>
   );
 }
 
-export default Schedule;
+export default SpeakerSchedulePage;
